refactor: co-locate thunder-gate unit tests with source files
Move test files from centralized src/__tests__/ to co-located __tests__/
subdirectories matching the thunder-develop convention. This improves
discoverability and maintainability as the app grows.
