+++ frontend/apps/thunder-gate/src/components/AcceptInvite/__tests__/AcceptInviteBox.test.tsx
@@ -19,7 +19,7 @@
 import {describe, it, expect, vi, beforeEach} from 'vitest';
 import {render, screen, fireEvent, waitFor} from '@thunder/test-utils';
 import userEvent from '@testing-library/user-event';
-import AcceptInviteBox from '../../../components/AcceptInvite/AcceptInviteBox';
+import AcceptInviteBox from '../AcceptInviteBox';
 
 // Mock useDesign
 const mockUseDesign = vi.fn();
